+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * st-search: Add filter to search
 * st-widget: Add settings button and draggable action as optional
+* st-tag-input: Add functionality to not allow user to type forbidden values
 
 **Breaking changes:**